Added updateAndFetchOneImpl() to share morecode
