+++ src/design-pattern-table-per-type/util/execution/index.ts
@@ -2,3 +2,4 @@ export * from "./fetch-one";
 export * from "./insert-and-fetch";
 export * from "./insert-row";
 export * from "./update-and-fetch-one-by-candidate-key";
+export * from "./update-and-fetch-one-impl";
